Other PR comment: Move check for lack of file providers into `CompilerCache` - `DefaultRazorViewEngineFileProviderAccessor.Value` now a `NullFileProvider`, not an empty `CompositeFileProvider`
